Remove duplicate class
